fix: 매우 큰 입력도 받을 수 있게 변경

+++ src/main/java/calculator/OutputView.java
@@ -1,9 +1,10 @@
 package calculator;
 
 import camp.nextstep.edu.missionutils.Console;
+import java.math.BigInteger;
 
 public class OutputView {
-    public void displayResult(int result) {
+    public void displayResult(BigInteger result) {
         System.out.println("결과 : " + result);
         Console.close();
     }
