+++ src/main/java/calculator/StringFacadeService.java
@@ -1,5 +1,7 @@
 package calculator;
 
+import java.math.BigInteger;
+
 public class StringFacadeService {
     private final StringCalculator stringCalculator;
 
@@ -8,7 +10,7 @@ public class StringFacadeService {
     }
 
     // 메서드의 실행 순서가 중요하므로 Facade 패턴을 적용.
-    public int addNumbersFromInput(String input) {
+    public BigInteger addNumbersFromInput(String input) {
         String[] delimiterAndNumbers = stringCalculator.addCustomDelimiter(input);
         return stringCalculator.addString(delimiterAndNumbers);
     }
